requirements.txt
numpy==1.11.1
pandas==0.18.1
DataSpyre==0.2.6
bokeh==0.12.3

